+++ src/interfaces/Description.ts
@@ -1,0 +1,5 @@
+interface Description {
+    name: string;
+    details: string;
+    [key: string]: string;
+}
